Files added via upload
In this task I had to use microsofts z3py theorem prover to satisfy constraints contained within the snippet file. The checksum I worked out in two stages, finding the correct bytes, and then finding the input to get the bytes, as I couldn't get the If statements to work correctly.